+++ R/LearnerAvg.R
@@ -32,7 +32,7 @@
 #'  * `optimizer` :: [`Optimizer`][bbotk::Optimizer] | `character(1)`\cr
 #'    [`Optimizer`][bbotk::Optimizer] used to find optimal thresholds.
 #'    If `character`, converts to [`Optimizer`][bbotk::Optimizer]
-#'    via [`opt`][bbotk::opt]. Initialized to [`OptimizerNLoptr`][bbotk::OptimizerNLoptr].
+#'    via [`opt`][bbotk::opt]. Initialized to `OptimizerNLoptr`.
 #'    Nloptr hyperparameters are initialized to `xtol_rel = 1e-8`, `algorithm = "NLOPT_LN_COBYLA"`
 #'    and equal initial weights for each learner.
 #'    For more fine-grained control, it is recommended to supply a instantiated [`Optimizer`][bbotk::Optimizer].
@@ -191,7 +191,7 @@ optimize_weights_learneravg = function(self, task, n_weights, data) {
       optimizer = pars$optimizer
       if (inherits(optimizer, "character")) {
         optimizer = bbotk::opt(optimizer)
-        if (inherits(optimizer, "OptimizerNLoptr")) {
+        if (inherits(optimizer, "OptimizerNLoptr") || inherits(optimizer, "OptimizerBatchNLoptr")) {
           optimizer$param_set$values = list(xtol_rel = 1e-8, algorithm = "NLOPT_LN_COBYLA", start_values = "center")
         }
       }